Chnaged project structure
Used routes in different file to simplify
Used json file to store datas
